Add a scale of spacing variables which can be used for spacing
This is just a suggestion and can be tweaked for personal
preference. One trick to achieving consistent visual spacing is
to use a spacing scale instead of just choosing random values.
My personal preference is to use Fibonacci numbers for the scale,
so this adds some custom CSS properties which map to Fibonacci
values. These values can then be used within the CSS code.

When these values are used in the entirety of the webpage (including
the margins for headings, paragraphs, etc.), then it really helps
to achieve visual harmony. That would be a bit too large of a
change for a single commit, so this only tweaks a few CSS rules to
use the new spacing scale

This also adds a few utility classes for adding a `margin-top` to
certain elements. This change is _very_ controversial, because many
developers do not like utility classes. However, I've recently
begun to question that opinion. In my opinion, the spacing should
be a first class citizen of the CSS container components into
which we place our components. However, there are a lot of instances
where we need something like "Glue Code" to bring different
components together within a UI and give them the correct amount
of spacing. I think that utility classes are a good use case for
this particular problem -- like for the `Copy 🍪 to clipboard`
button: in this instance the button needs a bit of a margin-top,
but if we were to write a whole new class for this specific
instance of the button within the UI, it would be extremely hard
to reuse it. With utility classes, it's easier to achieve a
_certain_ amount of reusability.

Note: This method of defining spacing has a downside of not working
in super old browsers (e.g. IE11) because they do not support CSS
properties. That is a tradeoff that I personally make for my own
personal site, but if that is problematic for you, then you could
also decide not to use CSS properties but just use the scale values
directly in the CSS code (e.g. `margin-bottom: 1.25rem` instead of
`margin-bottom: var(--spacer-md);`).